# apps/backend/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"DLocalWebhookrRouteBody": {
				"properties": {
					"payment_id": {
						"type": "string"
					}
				},
				"required": [
					"payment_id"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@revendiste/backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/webhooks/dlocal": {
			"post": {
				"operationId": "HandleDLocalWebhook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"received": {
											"type": "boolean"
										}
									},
									"required": [
										"received"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Receives payment status notifications from dLocal",
				"summary": "dLocal payment webhook",
				"tags": [
					"Webhooks"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DLocalWebhookrRouteBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}